requirements.txt
boto3==1.13.14
botocore==1.16.14
docutils==0.15.2
et-xmlfile==1.0.1
jdcal==1.4.1
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
openpyxl==3.0.3
python-dateutil==2.8.1
PyYAML==5.3.1
s3transfer==0.3.3
six==1.14.0
urllib3==1.25.9
